Port tests for Maybe to ava
